Split the oversized rollout test modules along cohesive seams (#312)
`test_typos_rollout_refresh.py` (448 lines) and
`test_typos_rollout_properties.py` (407 lines) both exceeded the
400-line ceiling. The refresh suite keeps cache-freshness and validator
metadata; the redirect policy, body-limit, and degradation-counter
tests move to `test_typos_rollout_degradation.py`. The properties suite
keeps the merge algebra and word-mapping laws; the degradation-counter
and remote-freshness properties move to
`test_typos_rollout_freshness_properties.py`.
